added database initializing function to copy over and rename all the auto generated test files
