+++ pyMAP/data/instrument/IBEX_lo_ETU/directEvents/load.py
@@ -1,6 +1,26 @@
 import numpy as np
 import pandas as pd
 
+def cal_headder(fil):
+    stuff = ''
+    for t in open(fil).readlines():
+        if '#' in t:
+            stuff+=t
+    head = []
+    for s in stuff.split('Group')[1].split('\n'):
+        sml = s.strip().strip('#').strip()
+        if '.' in sml:
+            try:
+                hnam = sml.split('.')[1].strip(')').strip('"')
+                if hnam not in head:
+                    head.append(hnam)
+                else:
+                    head.append(hnam+'2')
+            except:
+                print('dat import failed: %s'%fil)
+                return
+    return(head)
+
 def load(fil,
             use_filt = ['TOF0','TOF1','TOF2','TOF3'],
             filt_triples = False,
